Stop using the deprecated loop parameter in asyncio gather calls (#4)
* Stop using the deprecated loop parameter in asyncio gather calls

The feature was deprecated in py3.8 and finally removed in py3.10.
Tested the modified code to work with python3.11 & python3.8. for a
basic single version call.

* Drop useless loop assignment not used

Found by flake8. There's another finding but it's not related to this
change so didn't touch it.

* Add py3.11 to tox runs

Misc changes needed for tox to actually work:
- Add tox to test requirements
- Update flake8 repo address

* Cosmetic re-arrangements from tox run

* Update pre-commit env to get to a newer version of black

The previous version was incompatible with the used version of click.

* More cleanup to make black & flake8 happier

- Most changes are just cosmetic to the latest black spec
- Switch type testing to isinstance()
- Ignore E501 in test parsing & search since the test data is cleaner
as-is than forced to 79 chars.

---------

Co-authored-by: Jinna Kiisuo <jinna+git@nocturnal.fi>